Now including MTF
We're using the slanted edge of the resolution phantom to calculate the
average edge response, from this the line spread function and from this
the MTF.
For the moment this is only calculated for two images.
For the report, we have to add several MTFs to have a strong argument.